feat(backend/advertiser-service): 完善广告服务

+++ backend/advertiser-service/src/advertiser/advertiser.service.ts
@@ -1,8 +1,14 @@
+import { Pagination } from "@dongjiang-recruitment/nest-common/dist/decorator";
 import {
+  EntityPropertyNotFoundError,
   InjectRepository,
   Repository,
 } from "@dongjiang-recruitment/nest-common/dist/typeorm";
-import { Injectable } from "@nestjs/common";
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from "@nestjs/common";
 import { CreateAdvertiserDto } from "./dto/create-advertiser.dto";
 import { UpdateAdvertiserDto } from "./dto/update-advertiser.dto";
 import { AdvertiserInformation } from "./entities/advertiser.entity";
@@ -18,31 +24,43 @@ export class AdvertiserService {
     return await this.advertiserRepository.insert(createAdvertiserDto);
   }
 
-  async findAll() {
-    return await this.advertiserRepository.find({
-      skip: 1,
-      take: 5,
-    });
+  async findAll({ page, size, sort }: Pagination) {
+    try {
+      return await this.advertiserRepository.find({
+        skip: page * size,
+        take: size,
+        order: Object.fromEntries(sort),
+      });
+    } catch (error) {
+      if (error instanceof EntityPropertyNotFoundError) {
+        throw new BadRequestException(error.message);
+      }
+      throw error;
+    }
   }
 
   async findOne(id: string) {
-    return await this.advertiserRepository.findOne({
+    const advertiserInformation = await this.advertiserRepository.findOne({
       where: {
         advertiserInformationId: id,
       },
     });
+    if (!advertiserInformation) throw new NotFoundException();
+    return advertiserInformation;
   }
 
   async update(id: string, updateAdvertiserDto: UpdateAdvertiserDto) {
-    await this.advertiserRepository.update(id, {
+    const { affected } = await this.advertiserRepository.update(id, {
       ...updateAdvertiserDto,
       advertiserInformationId: id,
     });
+    if (!affected) throw new NotFoundException();
     return await this.findOne(id);
   }
 
   async remove(id: string) {
     const removed = await this.findOne(id);
+    if (!removed) throw new NotFoundException();
     await this.advertiserRepository.delete(id);
     return removed;
   }
